Added MuControllerBase class
