+++ src/MuHub.Api/Controllers/ErrorController.cs
@@ -10,63 +10,19 @@ namespace MuHub.Api.Controllers;
 /// <summary>
 /// Error controller that is used to handle all unhandled exceptions.
 /// </summary>
-public class ErrorController : ControllerBase
+public class ErrorController : MuControllerBase
 {
     /// <summary>
     /// Handles all unhandled exceptions.
     /// </summary>
-    /// <returns></returns>
+    /// <returns>An instance for <see cref="IActionResult"/>.</returns>
     [Route("/error")]
     [ApiExplorerSettings(IgnoreApi = true)]
     public IActionResult Error()
     {
         var exception = HttpContext.Features.Get<IExceptionHandlerFeature>()?.Error;
 
-        if (exception is DomainException domainException)
-        {
-            if (exception is ModelValidationException validationException)
-            {
-                return ValidationProblem(validationException);
-            }
-            
-            return Problem(domainException);
-        }
-
-        return Problem();
-    }
-    
-    private IActionResult Problem(DomainException exception)
-    {
-        if (exception.AdditionalContext is not null)
-        {
-            HttpContext.Items[ProblemDetailsConstants.AdditionalContext] = exception.AdditionalContext;
-        }
-        
-        var statusCode = RetrieveStatusCode(exception);
-
-        return Problem(statusCode: statusCode, detail: exception.Details);
-    }
-    
-    private IActionResult ValidationProblem(ModelValidationException modelValidationException)
-    {
-        var modelStateDictionary = new ModelStateDictionary();
-                
-        foreach (var validationExceptionError in modelValidationException.Errors)
-        {
-            modelStateDictionary.AddModelError(validationExceptionError.Key, validationExceptionError.Value.First());
-        }
-                
-        return ValidationProblem(modelStateDictionary);
-    }
-    
-    private static int RetrieveStatusCode(Exception exception)
-    {
-        return exception switch
-        {
-            NotFoundException => StatusCodes.Status404NotFound,
-            ClientFailureException => StatusCodes.Status400BadRequest,
-            _ => StatusCodes.Status500InternalServerError
-        };
+        return Problem(exception);
     }
 }
