+++ src/MuHub.Api/Controllers/V2/InfoController.cs
@@ -7,7 +7,7 @@ namespace MuHub.Api.Controllers.V2;
 [ApiController]
 [ApiVersion(2.0)]
 [Route("api/v{version:apiVersion}/[controller]")]
-public class InfoController : ControllerBase
+public class InfoController : MuControllerBase
 {
     [HttpGet]
     public IActionResult Index()
